// src/app/shared/models/deleteModalObj.ts
export class DeleteModalObj{
    title: string | undefined;
    msg : string | undefined;
    icon: string | undefined;
    errorMsg: string | undefined;
    successMsg: string | undefined;
    confirmButtonText?: string | undefined;
    cancelButtonText?: string | undefined;
    type?: string | undefined; 
}